+++ NEWS.md
@@ -7,6 +7,9 @@
   warning if they fail to get a fresh copy). This replaces the previous 
   ability to "pin" a url.
 
+* New `pin_browse()` shows you the source of the pin, either on the internet
+  or on your local file system (#435).
+
 * New `pin_delete()` supersedes `pin_remove()`, and can delete multiple
   pins at once (#433).